Separate presentation and business logic from Result
